split poll protocol files and unit tests
